Refactor app SCSS modules to use shared files
