+++ src/apps/BouncingPhysicsApp.tsx
@@ -10,7 +10,7 @@ import { OrbitControls } from 'three/addons/controls/OrbitControls.js'
 
 import { LOCAL_STORAGE_KEYS } from '@/utils/constants'
 
-import styles from '@/apps/BouncingPhysicsApp.module.scss'
+import styles from '@/apps/StandardAppWithGrab.module.scss'
 
 export const displayName: string = 'Bouncing Physics'
