--- src/apps/DraggableCameraApp.module.scss
@@ -1,24 +1,0 @@
-.app {
-  position: relative;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  min-height: 100%;
-}
-
-.container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-
-  cursor: grab;
-
-  &.grabbing {
-    cursor: grabbing;
-  }
-}
