+++ src/apps/StandardApp.module.scss
